+++ python/MRzeroCore/simulation/main_pass.py
@@ -12,12 +12,25 @@ import numpy as np
 # has can extract all information it wants during simulation.
 
 
+def rigid_motion(voxel_pos, motion_func):
+    """Shape of returned tensor: events x voxels x 3"""
+    def voxel_motion(time):
+        # rot: events x 3 x 3, offset: events x 3
+        rot, offset = motion_func(time)
+        rot = rot.to(device=voxel_pos.device)
+        offset = offset.to(device=voxel_pos.device)
+        return torch.einsum("vi, eij -> evj", voxel_pos, rot) + offset[:, None, :]
+    
+    return voxel_motion
+
+
 def execute_graph(graph: Graph,
                   seq: Sequence,
                   data: SimData,
                   min_emitted_signal: float = 1e-2,
                   min_latent_signal: float = 1e-2,
                   print_progress: bool = True,
+                  return_mag_p: int | bool | None = None,
                   return_mag_z: int | bool | None = None,
                   ) -> torch.Tensor | list:
     """Calculate the signal of the sequence by executing the phase graph.
@@ -50,6 +63,14 @@ def execute_graph(graph: Graph,
         The longitudinal magnetisation of the specified or all repetition(s).
 
     """
+    # This is a function that maps time to voxel positions.
+    # If it is defined, motion is simulated, otherwise the static data.voxel_pos is used
+    t0 = 0
+    voxel_pos_func = data.voxel_motion
+    if voxel_pos_func is None and data.phantom_motion is not None:
+        voxel_pos_func = rigid_motion(data.voxel_pos, data.phantom_motion)
+
+
     if seq.normalized_grads:
         grad_scale = 1 / data.size
     else:
@@ -71,6 +92,7 @@ def execute_graph(graph: Graph,
     # Calculate kt_vec ourselves for autograd
     graph[0][0].kt_vec = torch.zeros(4, device=data.device)
 
+    mag_p = []
     mag_z = []
     for i, (dists, rep) in enumerate(zip(graph[1:], seq)):
         if print_progress:
@@ -85,7 +107,7 @@ def execute_graph(graph: Graph,
             B1 = data.B1.sum(0)
         else:
             assert shim_array.shape[0] == data.B1.shape[0]
-            shim = shim_array[:, 0] * torch.exp(1j * shim_array[:, 1])
+            shim = shim_array[:, 0] * torch.exp(-1j * shim_array[:, 1])
             B1 = (data.B1 * shim[:, None]).sum(0)
 
         angle = angle * B1.abs()
@@ -118,7 +140,8 @@ def execute_graph(graph: Graph,
                 raise ValueError(f"Unknown transform {ancestor[0]}")
 
         # shape: events x coils
-        rep_sig = torch.zeros(rep.event_count, coil_count,
+        adc = rep.adc_usage > 0
+        rep_sig = torch.zeros(adc.sum(), coil_count,
                               dtype=torch.cfloat, device=data.device)
 
         # shape: events x 4
@@ -134,6 +157,18 @@ def execute_graph(graph: Graph,
         # Use the same adc phase for all coils
         adc_rot = torch.exp(1j * rep.adc_phase).unsqueeze(1)
 
+        # Calculate the additional phase carried of voxels because of motion
+        motion_phase = 0
+        if voxel_pos_func is not None:
+            time = t0 + torch.cat([torch.zeros(1, device=data.device), trajectory[:, 3]])
+            # Shape: events x voxels x 3
+            voxel_traj = voxel_pos_func((time[:-1] + time[1:]) / 2) - data.voxel_pos[None, :, :]
+            # Shape: events x voxels
+            motion_phase = torch.einsum("evi, ei -> ev", voxel_traj, rep.gradm * grad_scale[None, :]).cumsum(0)
+        t0 += total_time
+
+        mag_p_rep = []
+        mag_p.append(mag_p_rep)
         mag_z_rep = []
         mag_z.append(mag_z_rep)
         for dist in dists:
@@ -148,6 +183,8 @@ def execute_graph(graph: Graph,
                 continue  # skip dists for which no ancestors were simulated
 
             dist.mag = sum([calc_mag(ancestor) for ancestor in ancestors])
+            if dist.dist_type == '+' and return_mag_p in [i, True]:
+                mag_p_rep.append(dist.mag)
             if dist.dist_type in ['z0', 'z'] and return_mag_z in [i, True]:
                 mag_z_rep.append(dist.mag)
 
@@ -169,7 +206,9 @@ def execute_graph(graph: Graph,
             k1[0, :] = dist.kt_vec[:3]
             k1[1:, :] = k2[:-1, :]
             # Integrate over each event to get b factor (lin. interp. grad)
-            b = 1/3 * dt * (k1**2 + k1*k2 + k2**2).sum(1)
+            # Gradients are in rotations / meter, but we need rad / meter,
+            # as integrating over exp(-ikr) assumes that kr is a phase in rad
+            b = 1/3 * (2 * torch.pi)**2 * dt * (k1**2 + k1*k2 + k2**2).sum(1)
             # shape: events x voxels
             diffusion = torch.exp(-1e-9 * data.D * torch.cumsum(b, 0)[:, None])
 
@@ -186,21 +225,27 @@ def execute_graph(graph: Graph,
             # just by switching 2pi * (pos @ grad) to 2pi * pos @ grad
 
             if dist.dist_type == '+' and dist.emitted_signal >= min_emitted_signal:
-                T2 = torch.exp(-trajectory[:, 3:] / torch.abs(data.T2))
-                T2dash = torch.exp(-torch.abs(dist_traj[:, 3:]
-                                              ) / torch.abs(data.T2dash))
+                adc_dist_traj = dist_traj[adc, :]
+                if isinstance(motion_phase, torch.Tensor):
+                    adc_motion_phase = motion_phase[adc, :]
+                else:
+                    adc_motion_phase = motion_phase
+
+                T2 = torch.exp(-trajectory[adc, 3:] / torch.abs(data.T2))
+                T2dash = torch.exp(-torch.abs(adc_dist_traj[:, 3:]) / torch.abs(data.T2dash))
                 rot = torch.exp(2j * np.pi * (
-                    (dist_traj[:, 3:] * data.B0) +
-                    (dist_traj[:, :3] @ data.voxel_pos.T)
+                    (adc_dist_traj[:, 3:] * data.B0) +
+                    (adc_dist_traj[:, :3] @ data.voxel_pos.T) +
+                    adc_motion_phase
                 ))
                 dephasing = data.dephasing_func(
-                    dist_traj[:, :3], data.nyquist)[:, None]
+                    adc_dist_traj[:, :3], data.nyquist)[:, None]
 
                 # shape: events x voxels
                 transverse_mag = (
                     # Add event dimension
                     1.41421356237 * dist.mag.unsqueeze(0)
-                    * rot * T2 * T2dash * diffusion * dephasing
+                    * rot * T2 * T2dash * diffusion[adc, :] * dephasing
                 )
 
                 # (events x voxels) @ (voxels x coils) = (events x coils)
@@ -208,8 +253,10 @@ def execute_graph(graph: Graph,
                 rep_sig += dist_signal
 
             if dist.dist_type == '+':
-                # Diffusion for whole trajectory + T2 relaxation
-                dist.mag = dist.mag * r2 * diffusion[-1, :]
+                # Diffusion for whole trajectory + T2 relaxation + final phase carried by motion
+                dist.mag = dist.mag * r2 * diffusion[-1, :] 
+                if isinstance(motion_phase, torch.Tensor):
+                    dist.mag = dist.mag * torch.exp(2j * np.pi * motion_phase[-1, :])
                 dist.kt_vec = dist_traj[-1]
             else:  # z or z0
                 k = torch.linalg.vector_norm(dist.kt_vec[:3])
@@ -218,7 +265,7 @@ def execute_graph(graph: Graph,
             if dist.dist_type == 'z0':
                 dist.mag = dist.mag + 1 - r1
 
-        rep_sig *= adc_rot
+        rep_sig *= adc_rot[adc]
 
         # Remove ancestors to save memory as we don't need them anymore.
         # When running with autograd this doesn't change memory consumption
@@ -232,11 +279,14 @@ def execute_graph(graph: Graph,
     if print_progress:
         print(" - done")
 
-    # Only return measured samples
-    measured = torch.cat([
-        sig[rep.adc_usage > 0, :] for sig, rep in zip(signal, seq)
-    ])
-    if return_mag_z is not None:
+    measured = torch.cat(signal)
+
+    if return_mag_p is not None:
+        if return_mag_z is not None:
+            return measured, mag_p, mag_z
+        else:
+            return measured, mag_p
+    elif return_mag_z is not None:
         return measured, mag_z
     else:
         return measured