+++ python/MRzeroCore/simulation/pre_pass.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+from warnings import warn
 import torch
 import numpy as np
 import matplotlib.pyplot as plt
@@ -76,6 +77,9 @@ def compute_graph_ext(
             torch.cos(angle),
             torch.sin(angle/2)**2
         ], dim=1).type(torch.float32)
+    
+    if any(rep.pulse.angle > 2*np.pi for rep in seq):
+        warn("Some flip angles are > 360°, inhomogeneities produced by extra rotations are ignored by the pre-pass B1 estimation")
 
     return Graph(_prepass.compute_graph(
         seq,
@@ -107,7 +111,8 @@ class Graph(list):
             y-position of a state in the scatter plot
         color : str
             Use one of ``['abs(mag)', 'phase(mag)', 'latent signal', 'signal',
-            'emitted signal']`` as color of a state in the scatter plot
+            'latent signal unormalized', 'emitted signal']``
+            as the color of a state in the scatter plot
         log_color : bool
             If true, use the logarithm of the chosen property for coloring
         """
@@ -123,8 +128,12 @@ class Graph(list):
                 value = state.latent_signal
             elif color == "signal":
                 value = state.signal
+            elif color == "latent signal unormalized":
+                value = state.latent_signal_unormalized
             elif color == "emitted signal":
                 value = state.emitted_signal
+            else:
+                raise AttributeError(f"Unknown property color={color}")
             if log_color:
                 value = np.log10(np.abs(value) + 1e-7)
             return value