+++ src/pre_pass.rs
@@ -47,6 +47,7 @@ pub struct Distribution {
     /// latent_signal metric: if you want to measure states with a minimum signal of x,
     /// you should simulate states with a minimum latent_signal of <= x
     pub latent_signal: f32,
+    pub latent_signal_unormalized: f32,
 }
 
 impl Distribution {
@@ -158,7 +159,9 @@ pub fn comp_graph(
 
                 // Integrating (dt omitted) over k²(t) = ((1-x)*k1 + x*k2)^2
                 // gives 1/3 * (k1^2 + k1*k2 + k2^2)
-                let b = 1.0 / 3.0
+                // k's are in rotations / meter but we need rad / m -> * 2pi
+                use std::f32::consts::TAU;
+                let b = 1.0 / 3.0 * TAU * TAU
                     * dt
                     * ((k1[0] * k1[0] + k1[0] * k2[0] + k2[0] * k2[0])
                         + (k1[1] * k1[1] + k1[1] * k2[1] + k2[1] * k2[1])
@@ -334,6 +337,7 @@ pub fn analyze_graph(graph: &mut Vec<Vec<RcDist>>) {
         for mut dist in rep.iter().map(|d| d.borrow_mut()) {
             // Own signal is latent_signal if larger than what children produce
             dist.latent_signal = f32::max(dist.latent_signal, dist.emitted_signal);
+            dist.latent_signal_unormalized = f32::max(dist.latent_signal_unormalized, dist.signal);
 
             let max_contrib = std::iter::once(dist.regrown_mag)
                 .chain(
@@ -348,7 +352,9 @@ pub fn analyze_graph(graph: &mut Vec<Vec<RcDist>>) {
             for anc in dist.ancestors.iter() {
                 let contrib = (anc.rot_mat_factor * anc.dist.borrow().mag).norm() / max_contrib;
                 let tmp = anc.dist.borrow().latent_signal;
+                let tmp_unormalized = anc.dist.borrow().latent_signal_unormalized;
                 anc.dist.borrow_mut().latent_signal = f32::max(tmp, dist.latent_signal * contrib);
+                anc.dist.borrow_mut().latent_signal_unormalized = f32::max(tmp_unormalized, dist.latent_signal_unormalized * contrib);
             }
         }
     }
@@ -369,7 +375,7 @@ impl RotMat {
         let phase = if angle >= 0.0 { phase } else { -phase };
         let angle = angle.abs() * 180.0 / std::f32::consts::PI;
 
-        let index = angle.floor() as usize;
+        let index = angle.floor() as usize % 360;
         let t = angle.fract();
 
         let v1 = &avg_b1_trig[index];